+++ frontend/src/models/email.ts
@@ -1,0 +1,7 @@
+interface EmailInfo {
+  address: string;
+  subject: string;
+  content: string;
+}
+
+export type { EmailInfo };
